+++ EZLogger/Handlers/AuthorHandler.vb
@@ -1,0 +1,13 @@
+﻿Imports System.Windows
+Imports System.Windows.Forms
+
+Namespace EZLogger.HostForms
+    Public Class AuthorHandler
+
+        Public Sub OnOpenAuthorFormClick()
+            Dim host As New ReportAuthorHost()
+            host.StartPosition = FormStartPosition.CenterScreen
+            host.Show()
+        End Sub
+    End Class
+End Namespace